requirements.txt
fastapi
uvicorn
python-dotenv
slowapi
numpy
scipy
cachetools
requests
pandas
google-generativeai
httpx
tenacity
ta
websockets
scikit-learn
matplotlib
regex